requirements.txt
# 核心框架
torch==2.7.1
torchvision==0.22.1
transformers==4.57.1
ultralytics==8.3.217

# web框架相关
fastapi==0.118.2
uvicorn[standard]==0.37.0
pydantic==2.12.0
starlette==0.48.0

# 认证和监控
prometheus_client==0.23.1
python_jose==3.5.0


